requirements.txt
discord.py~=1.2
Pillow~=6.0
aiosqlite~=0.10
APScheduler~=3.5
yoyo-migrations~=6.0
